requirements.txt
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asgiref==3.11.1
boto3==1.42.78
botocore==1.42.78
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
dj-database-url==3.1.2
Django==5.2.11
django-anymail==14.0
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-storages==1.14.6
djangorestframework==3.16.1
docstring_parser==0.17.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jmespath==1.1.0
lxml==6.0.2
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.9.1
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.32.5
resend==2.26.0
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
